RA-12 v.0  Имплементировать интерфейс Parcelable и передать объект Recipe
