+++ app/src/main/java/com/example/recipeapp/Domain/Ingredient.kt
@@ -1,0 +1,11 @@
+package com.example.aseducationalproject.Domain
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Ingredient(
+    val quantity: String,
+    val unitOfMeasure: String,
+    val description: String,
+) : Parcelable
