--- app/src/main/java/com/example/recipeapp/Ingredient.kt
@@ -1,7 +1,0 @@
-package com.example.recipeapp
-
-data class Ingredient(
-    val quantity: String,
-    val unitOfMeasure: String,
-    val description: String,
-)
