--- app/src/main/java/com/example/recipeapp/Recipe.kt
@@ -1,9 +1,0 @@
-package com.example.recipeapp
-
-data class Recipe(
-    val id: Int,
-    val title: String,
-    val ingredients: List<Ingredient>,
-    val method: List<String>,
-    val imageUrl: String,
-)
