--- app/src/main/java/com/example/recipeapp/RecipeFragment.kt
@@ -1,26 +1,0 @@
-package com.example.recipeapp
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.recipeapp.databinding.FragmentRecipeBinding
-import com.example.recipeapp.databinding.FragmentRecipesListBinding
-
-class RecipeFragment:Fragment(R.layout.fragment_recipe) {
-
-//    class RecipeFragment : Fragment() {
-    private var _binding: FragmentRecipeBinding? = null
-    private val binding
-        get() = _binding
-            ?: throw IllegalStateException("Binding for ActivityMainBinding must be not null")
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?): View {
-        _binding = FragmentRecipeBinding.inflate(inflater,container,false)
-        return binding.root
-    }
-}
