+++ app/src/main/java/com/example/recipeapp/Recipes/RecipeFragment.kt
@@ -1,0 +1,45 @@
+package com.example.recipeapp.Recipes
+
+import android.os.Build
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.recipeapp.Categories.CategoriesListFragment.Companion.ARG_RECIPE
+import com.example.recipeapp.Domain.Recipe
+import com.example.recipeapp.R
+import com.example.recipeapp.databinding.FragmentRecipeBinding
+
+class RecipeFragment: Fragment(R.layout.fragment_recipe) {
+
+//    class RecipeFragment : Fragment() {
+    private var _binding: FragmentRecipeBinding? = null
+    private val binding
+        get() = _binding
+            ?: throw IllegalStateException("Binding for ActivityMainBinding must be not null")
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?): View {
+        _binding = FragmentRecipeBinding.inflate(inflater,container,false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val recipe: Recipe? = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            requireArguments().getParcelable(ARG_RECIPE, Recipe::class.java)
+        } else {
+            requireArguments().getParcelable(ARG_RECIPE)
+        }
+
+        binding.tvItemRecipe.text = recipe?.title
+
+
+    }
+
+
+
+}
