+++ app/src/main/java/com/example/recipeapp/Recipes/RecipesListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.recipeapp
+package com.example.recipeapp.Recipes
 
 
 import android.graphics.drawable.Drawable
@@ -7,12 +7,18 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
 import androidx.fragment.app.replace
-import com.example.recipeapp.CategoriesListFragment.Companion.ARG_CATEGORY_IMAGE_URL
+import com.example.recipeapp.Categories.CategoriesListFragment
+import com.example.recipeapp.Categories.CategoriesListFragment.Companion.ARG_CATEGORY_IMAGE_URL
+import com.example.recipeapp.Categories.CategoriesListFragment.Companion.ARG_RECIPE
+import com.example.recipeapp.DataTest.STUB
+import com.example.recipeapp.DataTest.STUB.getRecipeById
+import com.example.recipeapp.Domain.Recipe
+import com.example.recipeapp.R
 import com.example.recipeapp.databinding.FragmentRecipesListBinding
-import com.example.recipeapp.STUB
 
 
 class RecipesListFragment : Fragment(R.layout.fragment_recipes_list) {
@@ -75,9 +81,14 @@ class RecipesListFragment : Fragment(R.layout.fragment_recipes_list) {
     }
 
     fun openRecipesByRecipeId(recipeId: Int) {
+
+        //
+        val recipe: Recipe = getRecipeById(recipeId)
+        val bundle = bundleOf(ARG_RECIPE to recipe)
+        //
         parentFragmentManager?.commit {
             setReorderingAllowed(true)
-            replace<RecipeFragment>(R.id.mainContainer)
+            replace<RecipeFragment>(R.id.mainContainer, args = bundle)
             addToBackStack(null)
         }
     }
