+++ app/src/main/res/layout/fragment_recipe.xml
@@ -3,6 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 <TextView
+    android:id="@+id/tv_ItemRecipe"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:text="Рецепт"></TextView>
