+++ app/src/main/res/values/strings.xml
@@ -3,7 +3,6 @@
     <string name="category_text" translatable="false">Категории</string>
     <string name="favorite_text">Избранное</string>
     <string name="hello_blank_fragment">Hello blank fragment</string>
-    
     <string name="burger_item_category">"Бургеры"</string>
     <string name="item_category_description">Рецепты всех\nпопулярных видов\nбургеров</string>
     <string name="layout_category">Категории</string>
@@ -13,4 +12,5 @@
     <string name="image_rcipt_card">Image recipt card</string>
     <string name="recycle_view">Recycle view</string>
     <string name="recipe_image_description"></string>
+    <string name="recipe_image_content_description">%1$s сategory image</string>
 </resources>
